Part of the skill table is updated for spells to use proper cast times, still have to finish most of the remort stuff. Cast now gives feedback (player and room) and prevents casting if already casting.

+++ src/ack.h
@@ -1161,7 +1161,7 @@ struct skill_type
    short *pgsn;  /* Pointer to associated gsn    */
    short slot;   /* Slot for #OBJECT loading     */
    short min_mana;  /* Minimum mana used            */
-   short beats;  /* Waiting time after use       */
+   float beats;  /* Waiting time after use       */
    char *noun_damage;   /* Damage message               */
    char *msg_off; /* Wear off message             */
    char *room_off;   /* Wear off msg TO_ROOM    */
